requirements.txt
alabaster==0.7.16
Babel==2.14.0
certifi==2024.2.2
charset-normalizer==3.3.2
commonmark==0.9.1
docutils==0.20.1
exceptiongroup==1.2.0
idna==3.6
imagesize==1.4.1
importlib_metadata==7.0.2
iniconfig==2.0.0
Jinja2==3.1.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdit-py-plugins==0.4.0
mdurl==0.1.2
myst-parser==2.0.0
packaging==23.2
pluggy==1.4.0
Pygments==2.17.2
pytest==8.0.2
PyYAML==6.0.1
recommonmark==0.7.1
requests==2.31.0
snowballstemmer==2.2.0
Sphinx==7.2.6
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
tomli==2.0.1
urllib3==2.2.1
zipp==3.17.0
